+++ backend/app/data_products/output_ports/service.py
@@ -170,13 +170,6 @@ class OutputPortService:
 
         output_port.rolled_up_tags = rolled_up_tags
 
-        if not output_port.lifecycle:
-            default_lifecycle = self.db.scalar(
-                select(DataProductLifeCycleModel).where(
-                    DataProductLifeCycleModel.is_default
-                )
-            )
-            output_port.lifecycle = default_lifecycle
         if not self.is_visible_to_user(output_port, user):
             raise HTTPException(
                 status_code=status.HTTP_403_FORBIDDEN,
@@ -355,6 +348,12 @@ class OutputPortService:
         output_port_schema["data_product_id"] = data_product_id
         tags = self._fetch_tags(output_port_schema.pop("tag_ids", []))
         _ = output_port_schema.pop("owners", [])
+        if output_port_schema.get("lifecycle_id") is None:
+            output_port_schema["lifecycle_id"] = self.db.scalar(
+                select(DataProductLifeCycleModel.id).where(
+                    DataProductLifeCycleModel.is_default
+                )
+            )
         model = OutputPortModel(**output_port_schema, tags=tags)
 
         self.db.add(model)